requirements.txt
PyQt5>=5.8.2
Pillow>=4.1.1